'webkit-playsinline' not works well in some ios
My test device is iphone6,IOS10+. The 'webkit-playsinline' doesn't work on any browsers(Safari,chrome or some app webview).At the mean while, the 'playsinline' is totally suppoted and works well. So,I add the 'playsinline' when the option.playsinline is true. In case of some special needs,still keeping the 'webkit-playsinline'.